Вынес рейтинг в отдельный компонент
