+++ src/components/form-send-comment/form-send-comment.tsx
@@ -1,8 +1,10 @@
-import {ChangeEvent, useState} from 'react';
+import {useState} from 'react';
+
 import type {FormEvent} from 'react';
 import {LengthComment} from '../../const';
 import {fetchComments, sendComment} from '../../services/api-actions';
 import { useAppDispatch } from '../../hooks/use-store';
+import { RatingComponent } from '../rating-component';
 
 type PropsFormComment = {
     id: string | undefined;
@@ -16,10 +18,6 @@ function FormSendComment ({id}: PropsFormComment): JSX.Element {
 
   const isCommentLengthValid = !(comment.length >= LengthComment.MIN && comment.length <= LengthComment.MAX);
   const blockButton = isCommentLengthValid || rating === 0;
-
-  const handleChecked = (evt: ChangeEvent<HTMLInputElement>) => {
-    setRating(Number(evt.target.value));
-  };
 
   function onClickButtonSent(evt: FormEvent<HTMLFormElement>) {
     evt.preventDefault();
@@ -41,103 +39,18 @@ function FormSendComment ({id}: PropsFormComment): JSX.Element {
     setRating(rating);
   }
 
+  function getRatingFromComponent(ratingFromComponent: number): void {
+
+    setRating(ratingFromComponent);
+  }
   return (
     <form className="reviews__form form" action="#" method="post" onSubmit={onClickButtonSent}>
       <label className="reviews__label form__label" htmlFor="review">
         Your review
       </label>
-      <div className="reviews__rating-form form__rating">
-        <input
-          className="form__rating-input visually-hidden"
-          name="rating"
-          value={5}
-          checked={rating === 5}
-          onChange={handleChecked}
-          id="5-stars"
-          type="radio"
-        />
-        <label
-          htmlFor="5-stars"
-          className="reviews__rating-label form__rating-label"
-          title="perfect"
-        >
-          <svg className="form__star-image" width={37} height={33}>
-            <use xlinkHref="#icon-star" />
-          </svg>
-        </label>
-        <input
-          className="form__rating-input visually-hidden"
-          name="rating"
-          value={4}
-          checked={rating === 4}
-          onChange={handleChecked}
-          id="4-stars"
-          type="radio"
-        />
-        <label
-          htmlFor="4-stars"
-          className="reviews__rating-label form__rating-label"
-          title="good"
-        >
-          <svg className="form__star-image" width={37} height={33}>
-            <use xlinkHref="#icon-star" />
-          </svg>
-        </label>
-        <input
-          className="form__rating-input visually-hidden"
-          name="rating"
-          value={3}
-          checked={rating === 3}
-          onChange={handleChecked}
-          id="3-stars"
-          type="radio"
-        />
-        <label
-          htmlFor="3-stars"
-          className="reviews__rating-label form__rating-label"
-          title="not bad"
-        >
-          <svg className="form__star-image" width={37} height={33}>
-            <use xlinkHref="#icon-star" />
-          </svg>
-        </label>
-        <input
-          className="form__rating-input visually-hidden"
-          name="rating"
-          value={2}
-          checked={rating === 2}
-          onChange={handleChecked}
-          id="2-stars"
-          type="radio"
-        />
-        <label
-          htmlFor="2-stars"
-          className="reviews__rating-label form__rating-label"
-          title="badly"
-        >
-          <svg className="form__star-image" width={37} height={33}>
-            <use xlinkHref="#icon-star" />
-          </svg>
-        </label>
-        <input
-          className="form__rating-input visually-hidden"
-          name="rating"
-          value={1}
-          checked={rating === 1}
-          onChange={handleChecked}
-          id="1-star"
-          type="radio"
-        />
-        <label
-          htmlFor="1-star"
-          className="reviews__rating-label form__rating-label"
-          title="terribly"
-        >
-          <svg className="form__star-image" width={37} height={33}>
-            <use xlinkHref="#icon-star" />
-          </svg>
-        </label>
-      </div>
+
+      <RatingComponent rating={rating} setRating={getRatingFromComponent}/>
+
       <textarea
         className="reviews__textarea form__textarea"
         id="review"
@@ -168,3 +81,4 @@ function FormSendComment ({id}: PropsFormComment): JSX.Element {
 
 
 export default FormSendComment;
+
